+++ src/components/Login/LoginCard.css
@@ -1,0 +1,20 @@
+.hdr {
+    background-color: #333435; 
+    color: #fff; 
+    padding: 20px; 
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1); 
+    text-align: center; 
+    font-size: 18px; 
+    font-family: 'Arial', sans-serif; 
+    line-height: 1.5; 
+    transition: background-color 0.3s, color 0.3s;
+}
+.hdr h1 {
+    margin: 0 auto;
+    color: #F1F445;
+}
+
+.hdr:hover {
+    background-color: #e74c3c; 
+    color: #fff; 
+}
