+++ src/components/Login/LoginCard.js
@@ -1,0 +1,17 @@
+import React from "react";
+import {Link} from 'react-router-dom'
+
+import './LoginCard.css';
+
+const LoginCard = () => {
+    return (
+        <div className="hdr">
+            <button></button>
+            <h1>Welcome, Fuck Davito Mark Grigsby Strazzerri the Third, of 
+                diabeto the ninth!</h1>
+
+        </div>
+    );
+};
+
+export default LoginCard;
